telecom-fault-api/requirements.txt
fastapi
uvicorn
pymongo
motor
python-dotenv
passlib[bcrypt]
python-jose
redis